Remove PrecompileResult and move OOGResult
