+++ lib/evm/precompiles/01-ecrecover.ts
@@ -1,9 +1,10 @@
 import BN = require('bn.js')
 import { setLengthLeft, setLengthRight, ecrecover, publicToAddress } from 'ethereumjs-util'
-import { PrecompileInput, PrecompileResult, OOGResult } from './types'
+import { PrecompileInput } from './types'
+import { OOGResult, ExecResult } from '../evm'
 const assert = require('assert')
 
-export default function(opts: PrecompileInput): PrecompileResult {
+export default function(opts: PrecompileInput): ExecResult {
   assert(opts.data)
 
   const gasUsed = new BN(opts._common.param('gasPrices', 'ecRecover'))
